ProcessList: fix quadratic process removal when scanning
WARN WARN WARN WARN

NOTE: that we replaced all uses of Vector_indexOf with
ProcessList_indexOf

WARN WARN WARN WARN

This commit changes ProcessList_scan to remove Processes by index, which
is known, instead of searching by pid.

Searching by pid is quadratic in ProcessList_scan because the process we
are searching for is always the last process in the vector (the scan
starts from the back of the vector).

This commit also adds a new ProcessList_indexOf function to search for
processes by pid. This is more efficient than Vector_indexOf as it does
not require a comparison function, which cannot be inlined. The now
unused Vector_indexOf and Process_pidCompare functions have been
removed.